Explicitly load Active Record associations related errors
